+++ docs/Changelog.rst
@@ -8,13 +8,19 @@ Changelog
 
 **General**
 
+  * Add `wait_for_completion` to Allocation and Replicas actions.  These will
+    use the client timeout, as set by default or `timeout_override`, to
+    determine how long to wait for timeout.  These are handled in batches of
+    indices for now.
   * Allow `timeout_override` option for all actions.  This allows for different
     timeout values per action.
-    
+
 **Bug Fixes**
 
   * Disallow use of `master_only` if multiple hosts are used. Fixes #615
     (untergeek)
+  * Fix an issue where arguments weren't being properly passed and populated.
+  * ForceMerge replaced Optimize in ES 2.1.0.
 
 4.0.0a9 (27 Apr 2016)
 ---------------------